requirements.txt
flask==3.0.0
python-dotenv==1.0.0
langchain==0.0.350
openai>=1.6.1
langchain-openai==0.0.2
